Reduce compile warnings on windows.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23709 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Math/avtBezierSegment.h
@@ -77,12 +77,12 @@ public:
 
     unsigned int degree() const 
     { 
-        return size()-1; 
+        return (unsigned int)size()-1; 
     }
 
     unsigned int order() const
     {
-        return size();
+        return (unsigned int)size();
     }
 
     avtVector firstV() const
